requirements.txt
Django == 4.2.*
djangorestframework
mysqlclient
djangorestframework-simplejwt
timedelta
pandas
pyyaml